requirements.txt
requests~=2.32.2
bs4~=0.0.2
lxml~=5.2.2
beautifulsoup4~=4.12.3
python-dateutil~=2.9.0.post0